refactor: extract common functions, streamline dependencies
