+++ common/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2021"
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[features]
+test = []
+
+[dependencies]
+identity_iota.workspace = true
+identity_stronghold.workspace = true
+rand = "0.8"
