+++ common/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod test_utils;
+
+use identity_stronghold::StrongholdStorage;
+
+pub enum JwkStorageWrapper {
+    Stronghold(StrongholdStorage),
+    PKCS11,
+}
