+++ consumer/README.md
@@ -1,0 +1,3 @@
+# consumer
+
+Resolves a range of DID methods (similar to [DIF Universal Resolver](https://github.com/decentralized-identity/universal-resolver)) using the resolver of [identity.rs](https://github.com/iotaledger/identity.rs).
