+++ did_jwk/Cargo.toml
@@ -8,12 +8,18 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+common = { path = "../common" }
 did-jwk-extern = { version = "0.1.1", package = "did-jwk" }
-#identity_iota.workspace = true
-identity_iota = { version = "1.1.0", features = ["memstore"] }
+identity_iota.workspace = true
+#identity_iota = { version = "1.1.0", features = ["memstore"] }
+# identity_iota = { git = "https://github.com/iotaledger/identity.rs.git", branch = "main", features = [
+#     "memstore",
+# ] }
 identity_stronghold.workspace = true
 iota-sdk.workspace = true
+serde_json = "1.0"
 ssi-dids = "0.1.1"
+ssi-jwk = "0.1.2"
 tokio.workspace = true
 
 [dev-dependencies]
