+++ did_key/src/consumer.rs
@@ -4,12 +4,10 @@ use identity_iota::did::{CoreDID, DID};
 use identity_iota::document::CoreDocument;
 use identity_iota::resolver::Resolver;
 
-/// https://github.com/decentralized-identity/did-key.rs
-
 pub async fn resolve_did_key(did: CoreDID) -> std::result::Result<CoreDocument, identity_iota::core::Error> {
     println!("Resolving DID: {}", did);
     let key = resolve(did.as_str()).unwrap();
-    println!("key: {:?}", key.fingerprint());
+    println!("key fingerprint: {:?}", key.fingerprint());
     let document = key.get_did_document(Config::default());
     println!("document: {:#?}", document);
     let document = CoreDocument::from_json(&document.to_json().unwrap());
@@ -34,13 +32,18 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn resolved_did_key() {
+    async fn resolves_did_key() {
         let did = "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL";
         let document = resolve_did(did).await.unwrap();
 
         assert_eq!(
             document.id(),
             "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
+        );
+
+        assert_eq!(
+            document.verification_method().first().unwrap().id().to_string(),
+            "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
         );
     }
 }
