+++ did_web/Cargo.toml
@@ -8,14 +8,28 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+common = { path = "../common" }
 did-web-extern = { version = "0.2.2", package = "did-web" }
 # TODO: https://github.com/iotaledger/identity.rs/issues/1299
-identity_iota = { git = "https://github.com/iotaledger/identity.rs.git", rev = "4635a47", features = [
-    "memstore",
-] }
+# identity_iota = { git = "https://github.com/iotaledger/identity.rs.git", rev = "4635a47", features = [
+#     "memstore",
+# ] }
+# identity_iota = { git = "https://github.com/iotaledger/identity.rs.git", branch = "main", features = [
+#     "memstore",
+# ] }
+identity_iota.workspace = true
+identity_stronghold.workspace = true
+iota-sdk.workspace = true
 serde_json = "1.0"
 ssi-dids = "0.1.1"
+ssi-jwk = "0.1.2"
 tokio.workspace = true
 
 [dev-dependencies]
+common = { path = "../common", features = ["test"] }
+# iota-crypto = { version = "0.23", default-features = false, features = [
+#     "ed25519",
+# ] }
+iota_stronghold = { version = "2.0" }
+rand = "0.8.5"
 wiremock = "0.5"
